Initial OOP Design
The program is started by the GameEditorContainer.

The GameEditorWindow accesses the GameEditor instance from the
GameEditorContainer.

The GameEditor and all of its 'components' are now instances, and not
static. This allows us to do a logical flow of operation as the program
begins (such as loading prefabs, and configuring initial settings).